[refactor] Remove vm DApps related actions.
