--- src/types/DApp.js
@@ -1,7 +1,0 @@
-// @flow
-
-export type DApp = {
-  name: string,
-  publicKey: string,
-  signature: string,
-}
